Teacher_Reference/TalkingAI/requirements.txt
langchain
langchain_groq
python-dotenv
asyncio
deepgram-sdk
PyAudio